id_str: Handle full / duplicate cases better
This change refactors the hashtable in id_str.h to move most of the
table lookup code into a STR_GetEntryIndex() function, which handles
the case where the hashtable is totally full better (as was fixed
earlier for some, but not all, functions).

This helper returns the index which either contains the given key,
or which would contain it, if it were inserted. If the index is
ID_STR_INVALID_INDEX, the table is full (i.e., the key _cannot_ be
inserted). You can check which case by checking if the corresponding
key in the table for index is NULL — if so, the entry is not present,
and you can add it immediately. If not, this is the index of that
element — there's no need to re-strcmp().

(Note that this is invalidated if the table is modified.)

As a part of this, there is now a warning on duplicate entries in
debug mode (hence the fixes in the previous commit). This will be
disabled in a lot of cases with the next commit.

One potental future optimisation here is to make sequential lookups of
the same key fast. Though I suspect it'd cause more problems than it'd
solve if the hashtable isn't excessively full, and the string
could change.