ck_var: Add %override
We may want, when making mods, to be able to include all of the default
variable values from an episode, without having to include them
ourselves in the mod (and keep them up-to-date).

This allows us to make a variable file which only includes the values we
wish to override, and then use '%override' instead of '%include' to
include the original 'EPISODE.CKx'. Any variables which already exist
will not be updated.

Note that there _are_ some tricks with this: %override doesn't nest
obviously. In particular, once you're in a file with %override enabled,
then _no_ variable which already exists can be replaced. Ideally, there
shouldn't be any duplicate variables anyway, so hopefully this won't be
a problem.