ck_play: Move MapMiscFlagsCheck to EPISODE.CKx
As part of the slow quest to move everything out of the episode struct,
now the MapMiscFlagsCheck function is treated like any other action
function. In order to use it without an action, there's a new %function
variable type, which effectively creates an alias for a function. No
%colfunction equivalent exists yet, though.

Note that:
- Functions can only be set to builtins and NULL,
- There's a bunch of code to autodetect this if it isn't there.
- It's actually okay to not have a MiscFlagsCheck function if your world
  map is boring enough.
- These are still stored in the function map, not the variable map.
  Maybe it's worth merging these. Maybe it's worth splitting the
  variables up into new hashtables per-type. Food for thought.
- (The variable name lives in the varArena, though.)