id_fs: Make the maximum filename length a #define
We hardcoded the maximum filename length in FS_AdjustExtension() to 16.
This is (a) small, and (b) annoying to find. While it wasn't a problem
with the original game's data file (after all, everything needs to work
on DOS), it's possible people will use longer names for mods. (I'd still
not suggest this — Omnispeak does have a DOS version, after all.)

Still, bump it to 24, and use it in FS_AdjustExtension, as we'll need it
for future filename buffers as well.

(Maybe it'd also be nice to have a better error, as CK_Cross_strscpy()
will quit with a pretty unintuitive one.)