id_us: Make filenames configurable in EPISODE.CKx
The CONFIG.CKx and SAVEGAMx.CKx files should have configurable names in
EPISODE.CKx.

Note that SAVEGAMx.CKx is special here: the filename has to be modified
to include the savegame number. The original game hardcoded the offset,
and replaced 'x'. This uses the wildcard '?', and looks for it in the
string.

(One future possibility would be to support multiple '?' characters, and
hence more that 9 savegames, but that's probably overkill, so we don't.)