id_vl_dos: Fix WaitVBL/DelayTics
The WaitVBL/DelayTics implementation was just waiting for VBlank to be
asserted by the card: it should wait for VBlank to specifically end.

This ensures that we don't rush through all the pending VBLs during a
single VBlank, and matches the original game.

Fixes a number of dialogs (the cache box, etc) from disappearing
instantly on DOS/EGA.