id_ca: Support maps with header offsets of 0
Omnispeak has always assumed that a map doesn't exist if its header
(the value in MAPHEAD which corresponds to a position in GAMEMAPS) is
<= 0. However, this hits two issues:

1. The map header is unsigned, so modern compilers may replace this with
   an == 0 check, and
2. Some mods (e.g. The CK Guy's Keen 5.5) use 0 as a valid offset (as
   GAMEMAPS doesn't have the 'TED5v1.0' header string), so the map
   doesn't load.

Keen just checks < 0, and notes that -1 is used for nonexistant maps.
Looking myself, a lot of tools write 0 anyway, but it's better for us
to try to load a bad map than to fail to load a good one, so make the
check != 0xFFFFFFFF (equivalent to -1).