ck_us_2: Improved QuickSave to match FITF
Add a QuickSave confirmation dialog to match Foray in the Forest, which
has a very nice one. This also changes the behaviour slightly, QuickLoad
and QuickSave will match the previously-used savegame, not the last one
(unless no other savegame has been used).

The confirmation dialog can be changed in settings, but we've definitely
reached the edge of how many settings fit on-screen, so will have to
expand the settings list at some point. (Support scrolling, perhaps?)