id_sd_sdl: Use native-endianness for audio output (fix #71)
The OPL emulators (and/or our own mixing code) output in (pretty obviously)
the native endianness of the machine, but for SDL 1.2 and SDL 2, the AUDIO_S16
format we use is hardcoded to be little-endian, and the native-endian format
is called AUDIO_S16SYS.

Fortunately, SDL3 has restored sanity to the world, and uses S16 for native
and specific S16LE/S16BE for the specific versions. So nothing needs fixing there.

Thanks to @BeWorld2018 for finding and debugging this.