CMake: Bump CMake version
CMake is getting really annoying with its desire to fail if a CMakeLists.txt
specifies a CMake version which it feels is "too old".

By specifying a "maximum version" as well, we can tell them all to shut up, and
compile Omnispeak anyway.

But this is a moronic decision on the part of the CMake folks, and I respect them
less for having made it.