ck_play: Implement Extra Vibbles cheat
Add the F10+V debug cheat to add extra VBLs per frame. This makes it easier to
debug issues with both different game performance (e.g., physics issues when
tics/frame > rf_minTics), but also with video versus game timing.

(The plan is to rework some of the timing code to allow WaitVBLs to actually
use vsync on some setups. We already do this in the EGA/DOS backend, but not
on other systems (where we rarely run at 70Hz). It'll also make supporting
the CRT Terminator easier, which allows us to read the current scanline and so
detect VBlank more quickly and reliably. And maybe we'll do an entirely-VSync
based timing mode like the juj's hacks of Keen and FitF.)