ck_game: Add storedemo (/DEMO) support
Keen has a strange /DEMO option, which is sort-of documented here[1],
originally used for "store demos" or "kiosks". The idea is to have
the game only play the first couple of levels, not have access to the
menu, and reset itself if Keen gets bored. It also adds a bonus Ctrl+S
shortcut to toggle music and sound effects.

The idea is you can leave the game running for people to try out, without
risking that they might actually complete it (and therefore not want to
buy it).

Here we implement it, largely as-is. (We still don't support F9, so that's
not disabled.) There are a few small things worth noting:
- Unlike some things on the internet suggest, this doesn't disable cheats.
- The Ctrl+S shortcut doesn't clear the keys down, so can toggle itself
  repeatedly, seemingly not doing anything.
- Re-enabling music with Ctrl+S only takes effect when a new level is started.
- Omnispeak kept the level state inside the game state, but it _shouldn't_ be
  cleared in CK_NewGame(). This has been hacked around.

We also add (though not properly) the "One moment" message when teleporting
to Korath III. It doesn't do some of the strange "hack the UI code to write
directly to the frontbuffer", nor is the "One moment" message moddable at
present.

[1]: https://keenwiki.shikadi.net/wiki/Patch:Demo_mode