id_rf: Don't hardcode buffer/screen sizes
The RF buffer and screen sizes are canonically set by some macros in id_rf.h:
RF_BUFFER_{WIDTH,HEIGHT}_TILES and RF_SCREEN_{WIDTH,HEIGHT}_TILES, but a there
are a number of hardcoded values (in tiles, pixels, and units) spread across
both across the Refresh Manager and the rest of the game code (particularly
in ID_VL, the INTER/DEMO code, and the camera handling code in CK_PLAY).

Consolidate these into just the above macros. This does make it possible to,
for instance, run the game at different resolutions by hacking those values,
alongside the width/height in id_vl_private.h. This is still very buggy,
and breaks the determinism of demo playback, and will crash if the level
is smaller than the buffer size (e.g., the high scores screen), but crashes
much less immediately than it did before. Note that the "active area" is
not updated, nor are any of the main camera calulations or UI code, so only
the top, left-hand corner actually works proeprly.

A proper way of handling different resolutions / aspect ratios would involve
updates to the game code (and possibly shadowing all of the state for DEMO
playbacck), the ability to change this at runtime (though maybe not completely,
in-level), and maybe even a different ID_RF implementation which uses the GPU.

But for now, the code is just a little bit better.

Signed-off-by: David Gow <david@ingeniumdigital.com>