id_vl_sdl2vk: Fix memory corruption in SetupSwapchain
Due to a typo, we weren't allocating enough memory for the list of
surface formats in VL_SDL2VK_SetupSwapchain(). Instead of multiplying
the size of a swapchain format by the number of formats, they were being
bitwise-AND-ed together. This, by definition, doesn't even give enough
memory for one swapchain format, never mind multiple.

This bug brought to you by the 7/'&' key being next to the 8/'*' key on
the keyboard.