shaders: Don't hardcode the EGA size
We previously hardcoded the size of EGA mode 0Dh as 320×200 pixels
in the shaders used by the sdl2vk (Vulkan) and sdl3gpu (SDL GPU)
backends.

This is correct, but if we want to support arbitrary resolutions in the
future, we should probably keep the places this is hardcoded to a minimum.

With this change, higher resolutions seem to work.